Id removed from populate and add it to service. 5th query request ready. Finished: 1, 2, 4, 5

+++ m149_p2/src/main/java/gr/uoa/di/m149_p2/models/AbandonedVehicles.java
@@ -17,7 +17,7 @@ public class AbandonedVehicles extends Request{
     private Double daysParked;
     private String ssa;
 
-    public AbandonedVehicles(Integer id, Date creationDate, String status, Date completionDate, String serviceRequestNumber,
+    public AbandonedVehicles(Long id, Date creationDate, String status, Date completionDate, String serviceRequestNumber,
                              String typeOfServiceRequest, String streetAddress, Integer zipCode, BigDecimal x,
                              BigDecimal y, Integer ward, Integer policeDistrict, Integer communityArea,
                              BigDecimal latitude, BigDecimal longitude, String location, GeoJsonPoint point,
