+++ m149_p2/src/main/java/gr/uoa/di/m149_p2/models/Sanitation.java
@@ -13,7 +13,7 @@ public class Sanitation extends Request {
 
     private String codeViolation;
 
-    public Sanitation(Integer id, Date creationDate, String status, Date completionDate, String serviceRequestNumber, String typeOfServiceRequest, String streetAddress, Integer zipCode, BigDecimal x, BigDecimal y, Integer ward, Integer policeDistrict, Integer communityArea, BigDecimal latitude, BigDecimal longitude, String location, GeoJsonPoint point, Integer upVotes, String codeViolation) {
+    public Sanitation(Long id, Date creationDate, String status, Date completionDate, String serviceRequestNumber, String typeOfServiceRequest, String streetAddress, Integer zipCode, BigDecimal x, BigDecimal y, Integer ward, Integer policeDistrict, Integer communityArea, BigDecimal latitude, BigDecimal longitude, String location, GeoJsonPoint point, Integer upVotes, String codeViolation) {
         super(id, creationDate, status, completionDate, serviceRequestNumber, typeOfServiceRequest, streetAddress, zipCode, x, y, ward, policeDistrict, communityArea, latitude, longitude, location, point, upVotes);
         this.codeViolation = codeViolation;
     }
