+++ m149_p2/src/main/java/gr/uoa/di/m149_p2/models/TreeTrims.java
@@ -11,7 +11,7 @@ import java.util.Date;
 public class TreeTrims extends Request {
     private String treeLocation;
 
-    public TreeTrims(Integer id, Date creationDate, String status, Date completionDate, String serviceRequestNumber, String typeOfServiceRequest, String streetAddress, Integer zipCode, BigDecimal x, BigDecimal y, Integer ward, Integer policeDistrict, Integer communityArea, BigDecimal latitude, BigDecimal longitude, String location, GeoJsonPoint point, Integer upVotes, String treeLocation) {
+    public TreeTrims(Long id, Date creationDate, String status, Date completionDate, String serviceRequestNumber, String typeOfServiceRequest, String streetAddress, Integer zipCode, BigDecimal x, BigDecimal y, Integer ward, Integer policeDistrict, Integer communityArea, BigDecimal latitude, BigDecimal longitude, String location, GeoJsonPoint point, Integer upVotes, String treeLocation) {
         super(id, creationDate, status, completionDate, serviceRequestNumber, typeOfServiceRequest, streetAddress, zipCode, x, y, ward, policeDistrict, communityArea, latitude, longitude, location, point, upVotes);
         this.treeLocation = treeLocation;
     }
